#313 cleaned up WithNullValueOptionallyAllowedMixin and WithMaximumSelectionSize
