+++ superfields/src/main/java/org/vaadin/miki/superfields/gridselect/GridMultiSelect.java
@@ -8,7 +8,7 @@ import com.vaadin.flow.data.selection.SelectionEvent;
 import org.vaadin.miki.markers.WithHelperMixin;
 import org.vaadin.miki.markers.WithIdMixin;
 import org.vaadin.miki.markers.WithItemsMixin;
-import org.vaadin.miki.markers.WithMaximumSelectionSize;
+import org.vaadin.miki.markers.WithMaximumSelectionSizeMixin;
 import org.vaadin.miki.markers.WithValueMixin;
 
 import java.util.Collection;
@@ -23,7 +23,7 @@ import java.util.Set;
  */
 public class GridMultiSelect<V> extends AbstractGridSelect<V, Set<V>>
         implements WithIdMixin<GridMultiSelect<V>>, WithItemsMixin<V, GridMultiSelect<V>>,
-        WithMaximumSelectionSize<GridMultiSelect<V>>, WithHelperMixin<GridMultiSelect<V>>,
+        WithMaximumSelectionSizeMixin<GridMultiSelect<V>>, WithHelperMixin<GridMultiSelect<V>>,
         WithValueMixin<AbstractField.ComponentValueChangeEvent<CustomField<Set<V>>, Set<V>>, Set<V>, GridMultiSelect<V>> {
 
     private int maximumSelectionSize = UNLIMITED;
